Some Updates and News.
Okay so there's just a few changes to the actual code here. The
important thing is how to use this.
As Mike and some guys online pointed out to me, the Arduino IDE doesn't
handle dependencies between libraries very well. They haven't done a ton
to make it easy to create libraries or use object oriented programming
yet, especially with the YUN.

For now, the compiler won't complain if the CameraMod.cpp and
CameraMod.h files are in the arduino/libraries/ directory on your
machine.

I'm still having trouble getting the functions to work through a
separate program.ino, but I'll keep you guys updated.

~Zach